Enable idempotent deploys of workflows via the CLI
This PR utilizes the `UpsertWorkflow` endpoint, which idempotently
pushes workflow configuration to Inngest.

If the configuration is unchagned since the latest version the
config remains unchanged.

Otherwise, a new workflow version is created, by default as a draft.

This can push workflows live with the use of the `--live` flag.