Add support for scheduled funcs and URL-based funcs in `init`
This PR adjusts the initialization wizard to add:

- Schedule-based functions
- URL-based function invocations

When running through init we ask the following questions:

- Name of function
- Trigger (shcedule or event)
- Either the event or the cron schedule
- The function type (code or URL)
- Either the language or the URL.

When typing a cron schedule, we show a human format for the schedule
alongside the next time the function would run if immediately deployed.

URL-based functions use the HTTP runtime specified within each step.