Default to `test` workspace, with `--prod` flag and env var overrides
This PR removes all notions of "selected" workspaces, removing workspace
statefulness across the CLI.

We now default to the test workspace for any command (eg. deploying
functions, viewing events).  This can be overridden with a global
`--prod` flag, or by setting one of the following environment variables
to "production" (in order of priority):

- `ENV`
- `NODE_ENV`
- `ENVIRONMENT`