Change the starting type in state manager to `any`
This PR allows any data to initialize a new workflow, vs specifically a
map.  We also use this as an opportunity to enforce go 1.18 via the use
of `any`.

We use `any` vs generic types because generic types would force marking
_every_ user of state with generics (ie. the runner, the executor, etc).

The executor _does not_ care about starting data;  the state interface
handles this.  Therefore, `any` as a type param is the simplest and
easiest change.