Scaffold functions using steps
This PR ensures that we always `init` functions using a structure
suitable for step functions.  Instead of dumping a scaffold within the
root fn dir, we now create:

- `./events`, which contains event definitions using CUE types
- `./steps`, which holds the source for each step definition.

This allows us to create many steps for a single function without
complicating the root dir.

We also fix several issues relating to docker builds with multiple
steps, and tidy some logic for working with paths.