Add pause querying, pause iterator, and in-memory implementation
This commit adds an interface for querying pauses from the state store.
There may be many pauses returned when querying by events, so we add an
Iterator interface which allows backing stores to return events without
storing the entire set in memory.

We also provide an in-memory implementation of the query interface and
iterator, as well as a Redis implementation.

Finally, the dev server is updated to use the new interfaces for
resuming pauses from events.

A redis-specific note:

When a pause is set for a specific event, we store a hash of pauses by
ID -> pause (another copy) for each event.  This allows us to return an
iterator for reading all of the event's pauses.