Implement a basic queue interface
This commit introduces a basic queue interface.  The interface allows
us to enqueue steps to run at a specific time, and to provide a function
which should run when items are dequeued.

We change the in-memory queue to use this interface, and will provide
other implementations of the interface such as celery, SQS, etc. in a
future PR.