Add Server package for long running servers
This abstracts pre/start/stop across all components of the self-hosted
runner, ie. the API, the events API, and the executor.

We provide the ability for each component to specify its own pre-up
routines which, if errored, prevent the server from running.  This
allows us to eg. load secrets from each service prior to running.

We also handle interrupts and graceful shutdowns of the server.