Open-source pubsub mechanisms for the event stream
This PR open-sources pub/sub mechanisms for the event stream.  It allows
users to create publishers and subscribers to arbitrary messaging
systems (in-memory, nats, kafka, sqs, etc.), then handle running
functions every time an event is received.

We also provide the ability to subscribe to items in parallel from the
event stream, using a semaphore.  This grants us the ability to
dynamically change the concurrency rate based off of eg. memory pressure
in the future (ie. by changing the semaphore allocation for each event).

This backing implementation is used to buffer events from the event API
to our runner for scheduling queue executions.