Implement a Runner and Executor service
This commit implements two new services: a Runner and an Executor.

The runner is responsible for processing incoming events via pubsub.  It
enqueues new function runs for all functions which match incoming events
via the triggers.  It also resumes functions that have been "Paused",
where the pause matches the current event.  In both situations, the
runner pushes a new item into the backing queue which specifies the next
edge to run.  For new functions, this edge is the trigger. For paused
functions, it enqueues the outgoing edge stored within the pause.

The executor is responsible for reading items from the queue and
executing the item's step.  It stores the state from the step's action
and enqueues the next set of children, or manages errors when the step
fails.