Add edge tiemout handling to services
This commit introduces a new queue item for asynchronous edge timeouts.

When specifying functions, it's possible for you to specify that the
edge between steps should only be traversed if a matching event is
received within a time period.  This is easy:  we save a pause, and for
each event resume matching pauses.

You can also specify edges that must be traversed if this event is _not_
received.

This PR adds:

- A new Queue item to be executed when the edge's TTL occurs.
- A new Queue handler to check for the expired - but unconsumed - pause
  at TTL.  If present, we traverse the step via enqueueing the step.
- A state method for loading pauses by ID, including recently expired
  pauses.

We also include new test harness methods for the state harness.