Add initial, WIP implementation of event replaying
The code and output is all over the shop, but I have it working
end-to-end now.

From here, I can start smoothing out the output and make sure all of the
design decisions are sound.

I think the core logic can be a little bit simpler than it is right now,
so hopefully there's about the same amount of code as there is now with
the extra needed features adding (`--event-id`, etc).