Add `inngest run --snapshot` to return only event data
Replaces the existing `--event-only` flag and mostly mirrors its
functionality.

It differs in that it will now always return an array of events instead
of a single object (though the stdin reader now supports both).

It will now also return whatever events (and count) would be loaded for
the replay functionality.