Standardize parsing of cue overlay for windows
This commit changes cue overlay parsing such that it works on windows.
On windows, Cue requires a `C:\` prefix to all files loaded in the
overlay.  It also requires that path separators be OS-specific;
fs.WalkDir does not respect os-specific separators.

Finally, we change `inngest run` to exit when run errors, instead of
hanging.