Add support for .env files in local development
This commit introduces support for reading .env files for each function
when running locally via `inngest run` or `inngest dev`.  We do this by
creating a new `env` package which reads .env files for functions.

We extend the Docker driver with the ability to read from this env
reader, and we parse and set the reader when creating the execution
engine and drivers during `inngest run` and `inngest dev`.

Because .env files locally produce plaintext we do not reference
"secrets", here.  We'll extend the concept of secrets with eg. SOPS and
Vault integrations for encryption in the near future.