Basic non-partitioned redis queue implementation
This commit introduces a WIP basic non-partitioned queue implementation in
Redis, for people who desire a single backing store for the queue and state
together.

The queue works as follows:

- Each enqueued item gets a new ID
- We place the item into Redis, and enqueue a pointer to the ID in an
  ordered set with the visibility time as the score
- We pluck items from the ordered set when they're visible
- Each job is leased for a specific duration, and only worked on if the
  job is available.

The main downside to this approach is contention;  long running jobs
that aren't asynchronous will stay in the queue creating a backlog which
blocks other workers with multiple concurrent worker processes running.

This commit omits several crucial lua scripts for atomic transactions in
Redis.

**Future plans**:

In the future, we should move to a partitioned approach in which:

- We have a globally ordered set (zset) which stores function IDs and
  their earliest visibility time
- We have other zsets per function ID which store pointers to the actual
  queue item.
- We pluck function IDs from the globally ordered zset, lease the
  function ID per worker, then process each function ID as a separate worker
  until all jobs are done or we hit a processing limit.
- Once a worker is done with a function ID, it revokes its lease and
  sets the globally ordered score for the function to be the next
  earliest time in the queue, or nil.

This indirectly means that:

- All functions are processed in order of earliest job onwards
- Every function has an equal chance to be worked on, even taking into
  account backlogs
- We minimize contention because workers claim functions instead of
  individual job items