Add optional caching of expression data per pause
This lets us store a partial of all event/step/response data for async
edge expressions in the pause directly.

This reduces N+1 calls to load function state (which can be
costly/large) when evaluating edge expressions, as the data needed for
the expression is already returned within the pause.