Use custom slugging method for TS type names
We need to ensure that events with similar-but-different names are
appropriately separated.

For example, previously `github.pr` and `github/pr` would both become
`GithubPr`, which would then cause issues in TS.

With this, we now instead add a number of underscores based on the
character found. In our example above, `github.pr` becomes `Github_Pr`
and `github/pr` becomes` Github__Pr`.