Allow users to specify date strings within sleeps
This allows the SDK to calculate deltas based off of durations
themselves, vs passing the duration to our executor.  In effect, this
reduces latency:  there is additional processing time when a step
returns the sleep incurred by how we run steps via HTTP.

This helps decrease that latency to make sleeps more accurate.