Add event time to stream, plus change "replay" wording
`ts` seems to be empty with the dev server, which is an issue that needs
to be fixed.  We can and should also pull out the timestamp from the
ULID if timestamps are empty.

Events should use "replay" instead of retry, and functions should use
"rerun".  "Retry" is only used for failures :)