Allow queue kind -> queue name mapping
This PR introduces a new helper option for the redis queue, allowing
the queue to select queue names within `Enqueue` based off of job kinds.
If there's no mapping for a job kind the queue name is left nil, which
defaults to the workflow ID.

This helps resolve interface differences with the top-level queue
package and this state's implementation;  we currently can't enqueue a
redis_state.QueueItem directly and so can't choose a queue name without
this mapping.

In the future we should refactor the queue package and redis queue to
resolve differences and clean up this API.