Handle text responses in HTTP driver
We always need to account for text responses as there are situations out
of our control in which the SDK might not even be invoked.  Let's say
you deploy behind a proxy and the proxy is broken.  The proxy always
replies with text.  Our driver shouldn't error;  it should treat the
text as output and properly retry.