Peek partitions 1s ahead of time (#396)
This allows us to pull partitions with jobs available in 1s;  we can
then lease and consume jobs up to 1000ms before they're due to be ran
and wait in-memory prior to handlign the job.  This lowers latency for
jobs starting close to the tick of a second, improving p50-p99 latency.