Add basic function list view (#423)
* Add basic function list view

This adds a new view and GraphQL edge to query for registered functions. This also merges the loader on the dev server so the loader/readwriter that registers the functions is the same as the one that is used to fetch data in the API ensuring the memory is shared.

This also drops some legacy options for Docker that are no longer used. I don't believe this should break anything as this is just removing them from the dev server package only.

* Make id non-nullable

Co-authored-by: Jack Williams <jack@inngest.com>

* Replace string w/ enum to match

* Change function trigger to simpler iterable type

---------

Co-authored-by: Jack Williams <jack@inngest.com>