Add request version to metadata
This PR supports request versions per function run.  The request version
is discovered on the first SDK call with generator responses.  This
allows us to store the 'version' that an SDK uses for hashing/request
response styles.

We then use this version for the lifetime of the function.  Without
this, incompatible changes released are never backwards compatible -
breaking any in-progress functions.