refactor: use complete class names
Complete class names are safer to use with Tailwind CSS.

Tailwind can't find dynamically constructed class names and therefore will not generate the corresponding CSS, unless we always remember to add all possible permutations to the Tailwind config file.

If we forget to remove a class name in the config file, the class name will be unnecessarily added to the CSS file, even if it's not used anywhere.