Use client-side queries to decouple requests
This creates a better experience if a billing component fails with an error. The user should be able to see the error and it will be easier to debug in production. Additionally, this decouples the billable usage charge requests from the rest of the page. The billable usage can take a long time to load or it may fail in some instances which would previously cause the entire page to not load.