Add "Send to Dev Server" button to forward events
This adds a new button behind a feature flip that allows the user to quickly send the event to their dev server running on the default port on localhost. This saves the user from having to copy and paste the event and can speed up local replication of webhook events or events for debugging bugs using production events.

This also uses the new codeblock component which adds a necessary Copy button to the UI.