Add new queue kind to indicate a function to be ran.
The `start` kind represents a status that the function state has been
created but has not been worked on by the executor. Once the function
starts, subsequent executions will be changed to use the existing `edge`
kind.

There are essentially zero difference between `start` and `edge`. The
main reason to having this status is to be used as a differentiator
when introspecting the queue.