Add tick, allowing the dev server to respond faster (#1069)
By default, the dev server polls every 250ms and forwards the in-memory
state store every second.  This introduces a --tick parameter, allowing
the dev server to tick faster to match production.

This is not the default;  we want to save people's CPU during regular
development.  It is aimed at CI/testing only.