Remove inaccurate error return
This no-data code path was hit incorrectly and would produce false errors when data was loading. I debugged this a bit and didn't get to the bottom of this path, but I believe it's due to the asynchronous request within the hook and some sort of in-between state where the request should be sent, but the return path with !data was hit first preventing further execution.