Add GCRA to leasing queue items directly
This commit adds GCRA to queue items.  It's possible to specify
GCRA/throttling parameters directly when enqueueing an item.  If a queue
item contains GCRA paremters, the queue will apply GCRA when attempting
to lease the queue item.

This ensures that the queue atomically, but selectively, applies
throttling as it attempts to claim items to be processed.