Add forcing to partition items
This allows us to force a partition pointer to be kept at a future time,
regardless of enqueueing new jobs.

We do this to push out queue pointers in the case of concurrency limits;
if there's no capacity, we don't want to recheck the partition for at
least another second or two - certainly not many times per second.