Add GCRA to leasing queue items directly (#1263)
* Add GCRA to leasing queue items directly

This commit adds GCRA to queue items.  It's possible to specify
GCRA/throttling parameters directly when enqueueing an item.  If a queue
item contains GCRA paremters, the queue will apply GCRA when attempting
to lease the queue item.

This ensures that the queue atomically, but selectively, applies
throttling as it attempts to claim items to be processed.

* Add forcing to partition items

This allows us to force a partition pointer to be kept at a future time,
regardless of enqueueing new jobs.

We do this to push out queue pointers in the case of concurrency limits;
if there's no capacity, we don't want to recheck the partition for at
least another second or two - certainly not many times per second.

* Handle rate limit leasing errors in processor

* Ensure that non-sequential scanners randominze all available partitions

* Add burst tests