Create our own data structure for otel spans to workaround otel spec limitations (#1248)
The otel library is built on an assumption that span instrumentation API should be within a piece of local resource. In other words, you should not be able to reconstruct a remote span and modify it on another resource over the wire.

While that's a very reasonable and solid argument in most cases, abiding by that usage will make our trace representation incomprehensible for our users.
The goal of Inngest is to abstract the complexity of the inner workings of an event-driven system and make it as close as possible to a "normal" function call on the stack.

In order to achieve that objective for tracing, we'll need to propagate spans across network boundaries in a way that from the data's perspective, as if there were none. Therefore, we need a data structure that can
1. the propagation of span across pub/sub, queue systems
2. override certain span parameters to make it look as it is still the same span
3. keep the inner workings as transparent as possible while allowing multiple submission of the same span

The necessity for `#2` is mainly as a function run progresses, we need to continuously update the function run span's end time / duration. And by allowing multiple submission of the same span with different end times, we can dedup on our end to make sure the last one is always the correct one to stay.

That way, as the function run progresses, the data will also reflect the run in a way that will simulate a stack function call.
This data will also be the source to power the new UI coming up to allow better visibility of what's happening with user's workload.

### Assumptions
Dedupping of similar spans is out of scope, and is the responsibility of either the storage engine (e.g. Clickhouse) or something that's below the application layer.

---------
Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>