Add lease denylists (#1334)
This commit introduces lease denylists for queue items.  When attempting
to lease queue items, we pull a chunk of work from the queue in FIFO
order as a list of queue items. We iterate through this list, and
attempt to lease the queue item.

The queue item may fail to lease because of concurrency keys or
throttling keys.  In this case, we may still keep iterating on the list
to attempt to process *other* items with *different* concurrency keys.

We add these failed keys to a denylist, and never attempt to lease those
queue items within the same iterations.  This keeps FIFO order and
prevents queue churn.

Note that denylists are temporary and only exist per executor for the
function currently being peeked.