Include error message in history item output (#1402)
Currently, we do not properly enrich the history run item with errors included in the driver response. This causes step limit errors to vanish in the UI. This fix simply adds the error message to the history item output. Long-term, we should properly set and pass on the error code to be rendered nicely in the UI.

---------

Co-authored-by: Aaron Harper <aaron@inngest.com>