Make system partitions backwards compatible
Fixes INN-3496

This ensures that new deployments are able to read old system partitions which only provide a queueName.

All partitions with a queueName value set are automatically assumed to be system partitions (we don't use the override feature anywhere else).

Previously, we used another partition type, but that would never be set because we don't update existing partitions (HSETNX)