Fix button flicker when polling
Polling changes the state of the button too quickly and causes the UI to flicker unnecessarily. Disabling the button when polling doesn't matter too much as the view is always live and up to date. There is a potential better fix than this, but this is a simple improvement for situatons where we're polling (the dev server).

With the faster polling in #1724, this problem got even worse.