Fix external Redis issues when running in single node mode (self-hosting) (#2029)
* Parse Redis URI components into options. Fixes #2025.

The Redis URI was previously being directly used and was not properly parsed so options including username or passwords were not accepted. This now utilizes the library's connection string parsing to correctly accept various options.

* Fully disable snapshot imports and exports when using external Redis.

This builds on the fix in #1965 and potentially fully fixes #1916.

Prior to this, even with external redis, snapshots were made or attempted to be imported. If there is an external Redis setup, snapshots should never be used.

This also changes the method name to be more explict.

* Implement missing methods required by interface