Use the HTTP hostname in the returned WebSocket URL.
When the dev server, or self-host is running on something other than 127.0.0.1, like on a public URL or in a Docker container using a custom network hostname, the dev server should attempt to automatically re-write the websocket URL on the same hostname that the HTTP request was received on. This saves the end user from having to intercept and re-write the gateway endpoint.