Channel APIs (#2212)
* duration tracking

* Add support for chunked streaming via broadcasters and subscribers

This commit introduces an API for publishing data to channels/topics
within streams.

The API reads all data then publishes to the streaming channel by
default.  If the request's content type is "text/event-stream" or
"text/stream", the request body will be chunked and sent to the
broadcaster every ~400ms, allowing for fast UI updates.

* Implement chunked streaming

* Go initialisms

* Re-add run IDs for future debugging purposes

* Fix nits and lints

* UTC

* Update types

* json tag

* Use official SDK implementation for realtime test

* Lint