Ensure next handler is called when auth middleware is used.
In cloud, we pass auth middleware that loads keys so clients can fetch the token with a signing key (or future auth strategy). This ensures that the resulting middlware calls the next handler.

This was debugged and tested in the upstream cloud monorepo. This is not breaking as auth middleware is not used in OSS and it's broken in prod.