Make function run scheduling entirely idempotent (#2317)
Due to the lack of transactions in Redis, we need to make sure scheduling a function run with events is entirely idempotent end to end.

In the cases of crashes and reruns, this won't end up creating duplicated runs nor will it miss scheduling runs, and will have exactly once execution.

---------

Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>