Add batch size limit internally to make sure batches don't exceed a certain size (#2393)
This change is to make sure a run won't have a batch that exceeds a certain size in bytes.
Since batches are accumulation of events, and the size increases rapidly if each event has a large payload size.

Needed to update miniredis version so we can use the command `MEMORY USAGE` for checking size of the batch.

---------

Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>