Trust run IDs from the SDK when ingesting userland spans
This circumvents an issue whereby userland spans could be ingested
before the parent/root spans have been saved to the data store.

This means that the previous check that pulls the trace root from the DB
would find nothing, and therefore we intentionally fail when ingesting.

Because we don't want to change this method, we instead rely on the client
SDK to send the run ID and trust that it is correct.

This is not a great long-term solution, but post tracing refactor (where writing
trace spans are part of the critical workflow), this should not be an issue and
we can return to checking that runs exist and have the correct owner before
contributing userland spans to it.