Implement eager cancellation (#2474)
Currently when cancellation is initiated, we only do JIT cancellation.
With key queues introduced, we need a way to eagerly cancel items so we can clear out backlogs more efficiently.

This change introduces additional queue reader APIs to find queue items that are candidates for cancellation when one is initiated.

Change is made to allow it to be used with the current bulk cancellation API as well, and potentially the bulk cancellation initiated from the UI.

---------
Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>