Split HTTP and driver (#2473)
* Add realtime tee publish util

This PR adds a utility to tee readers into realtime, with additional
tests.

* close the body in tests

* Split HTTP and driver

This commit separates HTTP clients and comms from the "driver"
implementation.  We do this because:

- The driver executes steps
- And we implement HTTP requests in many places:
    * Gateways
    * Syncs / deploys
    * Pings

We should NOT use `httpdriver` as an HTTP requester and a step executor,
as this confuses logic.

Additionally, we move to use the `SerializabelRequest` type and clean up
requests/response types instead of using stdlib.

* Implement publish auth for realtime API, beyond API keys

* fix merge conflict

* Update httpstat

* Exports

* lints

* Add more metadata

* Unexpected end

* once again fix tests

* remove lite

* Add additional response fields