Fix custom concurrency active checks
Previously, the active check for custom concurrency would check

- the in progress key for the custom concurrency (correct)
- the ready queue for the backlog's shadow partition (wrong!)

The latter was an issue as the custom concurrency key may be scoped to an environment or account, so the key can be reused by multiple functions.

Example
- In progress key: {q:v1}:concurrency:custom:e:b6d7651b-...:1nt4mu0skse4a
- Active key: {q:v1}:v2:active:custom:e:b6d7651b-...:1nt4mu0skse4a

The canonical key ID (<scope>:<entity>:<hash>) can be used by multiple functions. This means the active keys will be shared for a key across the function boundaries (as is expected: If you declare a custom concurrency key on user ID on the environment level, you want all queue items for that user to be accounted for across the environment).

If you only look at the ready queue for the current backlog, this will exclude all other function ready queues that may be the actual ready queue.

Instead of using a static ready queue, we now use the same script as we use for account active keys to simply retrieve the queue item, check which partition it belongs to, and then comparing the in progress queue for the custom concurrency key, then the ready queue for the item's partition.

We can simplify the active checker by removing the extra script and reusing the existing logic for partition-level checks (which would work with a static ready and in progress queue based on the shadow partition).