First pass at EventsV2 dev server resolver (#2656)
The EventsV2 gql schema is copied directly from monorepo.

Currently unimplemented functionality:

- function runs
- source
- filtering on internal events
- cel query

This PR introduces changes to the cqrs methods and sqlc queries to support filtering on multiple event names and new queries for getting total count.
It's possible we will have to switch these to goqu to support additional filtering in the follow up work.

I am also punting on parity in Postgres to a follow up.

* Move TotalCount to a separate resolver

use GetFieldContext to get filter input from parent resolver/query

* Move EventV2.Raw to separate field resolver

There are a few issues with the current implementation
- There is an N+1 when we are retrieving Raw when querying EventsV2
  because currently this will query the events table again in GetEventByInternalID.
  This is probably not a big deal because we won't be querying for Raw
  under EventsV2 in the UI (only in the test for now)
- There is a single EventV2 resolver upcoming. Raw will definitely be
  queried under this from the UI. As currently implemented, we would
  make 2 of the same query to the DB to resolve the other fields + Raw.

Both of these can be mitigated by using a data loader instead in a follow up