Fix bug that kills the ws connection when polling happens fast
Sometimes executors will poll at a precise moment and get the response
faster than gRPC, what was happening in that case is that the lease gets
cleaned up by the executor. The gateway fails to find the executor IP to
get the right gRPC client and the ws connection gets killed because of
the error.