Add error reporting interface to logger (#2718)
Provide error reporting as part of the logger interface.

This helps encapsulate how errors are being reported without littering the codebase with a bunch of sentry deps.

Log attributes are also inherited into the error tags as much as possible as long as it can be rendered into a string, so it should simplify replacing error logs with error reports.

---------

Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>