Do not save event payload in `Span.triggers` (#2810)
We currently store the event message in the span twice, once in `Span.triggers` and once in `Span.events`. This PR stops populating the event in `Span.triggers` to avoid the duplication that causes the message size to blow up beyond what can flow through Kafka.

This should not cause other downstream issues as `Span.triggers` is not ingested into clickhouse in the first place. 

I verified that `inngest.customer_trace_runs.triggers` continues to be populated with this change by manually running a test and checking clickhouse entries for the test's runId:

<img width="2535" height="725" alt="Screenshot 2025-08-07 at 7 31 40 PM" src="https://github.com/user-attachments/assets/cbcfcb68-2c47-427c-9516-c00f8ab826ee" />