fix CQRS conversion issue with `copier` (#2868)
This addresses pitfall with using the `copier` lib, where it doesn't handle `sql.Null*` conversions correctly.

Instead of relying on reflection rabbit hole, we should just provide concrete converters to handle the conversions. This makes it less likely to have surprises.

Also added basic functional tests for a CQRS query using in-memory sqlite.
We'll need to add more coverage for this.


---------
Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>
Co-authored-by: Jakob Evangelista <79593812+jakobevangelista@users.noreply.github.com>