Add run claim JWTs for sync APIs
This allows us to create single-scoped JWTs that allow access to a
specific run ID's outputs.  Critically, we create these JWTs whenever a
sync fn checkpoints and turns async.  The sync fns can then take these
JWTs and give them to the original caller of the API, allowing them to
fetch the output of the async fn in the future.