Fix gcra for Lua 5.2 -> 5.3
Garnet uses KeraLua which embeds Lua version 5.4, whereas Valkey embeds
Lua 5.1. For Lua 5.3, floats are now represented with a trailing .0 when
converted to strings.

> The conversion of a float to a string now adds a .0 suffix to the
result if it looks like an integer. (For instance, the float 2.0 will be
printed as 2.0, not as 2.) You should always use an explicit format when
you need a specific format for numbers.

(Formally this is not an incompatibility, because Lua does not specify
how numbers are formatted as strings, but some programs assumed a
specific format.)

This impacts existing scripts which do not explicitly convert numbers to
strings. This change fixes the gcra() function which provided a float
value to a redis.call() function.

For more details, see https://www.lua.org/manual/5.3/manual.html#8